Types.hs -> Protocol.hs
